(waybridge) start adding xwayland support
XWayland works in 'normal' non-rootless mode but the thing itself is
rather buggy and don't respond to resize/configure hints. The mode
others are using to transparently support mapping X windows requires a
special compositing window manager to work. The normal tactic for that
is to pull it in as part of the compositor and have that spawn and spin
up XWayland instances. True to form we'll do this a bit differently and
do all the icky xcb- stuff in a separate 'arcan-xwayland-wm' binary /
process.