(shmif) server/proxycon tool works interactively
This should put us across the big main hurdle with the rest being slow
but steady mapping of features (descriptor transfers, accel.  handle
sharing, audio, subsegments, ...), but it is now possible to use the
'proxycon' test to do interactive MiM of a non-auth connection.

Add encoding and the parts that are already in netpipe and 'x-level'
of transparency should be here.